up: make setup persistent volumes a component
